Make refractoriness tests also run with C++ code generation, change return type of the int_ function to int

+++ brian2/codegen/functions/numpyfunctions.py
@@ -152,7 +152,7 @@ class IntFunction(Function):
 
     def code_cpp(self, language, var):
         support_code = '''
-        double int_(const bool value)
+        int int_(const bool value)
         {
 	        return value ? 1 : 0;
         }
